+++ CHANGELOG.md
@@ -12,7 +12,11 @@
 
 ### Enhancements
 
-* None.
+* Add new default `invisible_character` rule that detects invisible characters
+  like zero-width space (U+200B), zero-width non-joiner (U+200C),
+  and FEFF formatting character (U+FEFF) in string literals, which can cause hard-to-debug issues.  
+  [kapitoshka438](https://github.com/kapitoshka438)
+  [#6045](https://github.com/realm/SwiftLint/issues/6045)
 
 ### Bug Fixes